--- src/backup/backup.css
@@ -1,21 +1,0 @@
-* {
-    box-sizing: border-box;
-    overflow-y: scroll;
-    --webkit-overflow-scrolling: touch;
-}
-html {
-    overflow: hidden;
-}
-body {
-    /*overflow: auto;*/
-    margin: 0 !important;
-    font-family: 'Segoe UI', 'Helvetica', sans-serif;
-}
-body, #___gatsby, #___gatsby > *, .full-page {
-    width: 100%;
-    max-width: 100%;
-    max-height: 100%;
-    height: 100%;
-    overflow-y: scroll;
-    --webkit-overflow-scrolling: touch;
-}
