+++ src/components/renderers/textRenderer.js
@@ -1,20 +1,14 @@
 // @flow
 import * as React from 'react'
-import {
-  Card,
-  MarkdownContent,
-  PrimaryView,
-  Subtitle,
-  Title,
-} from '../redesign'
 import toTitleCase from '../../utils/toTitleCase'
+import { Card, MarkdownContent, PrimaryView, Subtitle, Title } from '../styles'
 
 type Props = {
   data: Object,
   html: string,
 }
 
-export default class MarkdownRenderer extends React.Component<Props> {
+export default class TextRenderer extends React.Component<Props> {
   render() {
     const { data, html } = this.props
     return (
