+++ src/pages/404.js
@@ -1,10 +1,13 @@
 // @flow
 import * as React from 'react'
 import { graphql } from 'gatsby'
-import Layout from '../components/redesignedLayout'
+import Layout from '../components/layout'
 import LowFundsAnimation from '../components/animations/lowFunds'
-import { ErrorPageHeader } from '../components/styles'
-import { ErrorPageLink } from '../components/redesign'
+import {
+  ErrorPageLink,
+  ErrorPageHeader,
+  ErrorPageDescription,
+} from '../components/styles'
 
 type Props = {
   data: Object,
@@ -19,13 +22,11 @@ export default class Frameworks extends React.Component<Props> {
             <div className="container">
               <div className="grid-row grid-row--middle grid-row--center">
                 <div className="grid-col-12 grid-col-6-md">
-                  <ErrorPageHeader className="u-font-maison-mono u-font-light">
-                    404
-                  </ErrorPageHeader>
-                  <h2 className="u-color-grey">
+                  <ErrorPageHeader>404</ErrorPageHeader>
+                  <ErrorPageDescription className="u-color-grey">
                     Sorry, we can&rsquo;t find the page you&rsquo;re
                     looking&nbsp;for
-                  </h2>
+                  </ErrorPageDescription>
                   <ErrorPageLink
                     to="/frameworks/"
                     className="o-button-text u-margin-top-large-md"
@@ -43,12 +44,12 @@ export default class Frameworks extends React.Component<Props> {
             <div className="container">
               <div className="grid-row grid-row--center">
                 <div className="grid-col-12 grid-col-6-md u-align-center-md">
-                  <h3 className="u-font-masion">Did you know&hellip;</h3>
+                  <h3>Did you know&hellip;</h3>
                   <p>
                     In 2008, there were 4,916 claims (totalling &pound;113,000)
                     for bank notes which had been eaten by pets.
                   </p>
-                  <p className="u-font-size-h1">🐶 🐱</p>
+                  <h1>🐶 🐱</h1>
                 </div>
               </div>
             </div>
